Fixed a Bug Regarding Second [Xe|La|Lua|Pdf]Tex Invocation
If we did BibTeX, we need to invoke [Xe|La|Lua|Pdf]Tex two times to get
a stable result. I tried to do that with

if ["$cycle" -ge 2] ; then
  ...

before, but that does not work as anticipated and returns

scripts/tex.sh: line 62: [1: command not found
scripts/tex.sh: line 68: [: missing `]'

or something like that.

We now do it via

if (("$cycle" > 1)) ; then
  ....
  
which seems to work.